Factor out creation of downstream environment
I'll surely need to use this more than just for proxy sync tests.
